feat: yearlyReviewDetailInfo function
